# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GVoice API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /channels/{channelId}/messages:
    get:
      summary: List channel messages
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: cursor
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200': { description: OK }
    post:
      summary: Send message
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 4000 }
      responses:
        '201': { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT